+++ CHANGELOG.md
@@ -10,11 +10,11 @@
 * [Racecar::Consumer] Allow providing `additional_config` for subscriptions
 * [Racecar::Consumer] Provide access to `producer` and `consumer`
 * [Racecar::Consumer] Enforce delivering messages with method `deliver!`
+* [Racecar::Consumer] instead of raising when a partition EOF is reached, the result can be queried through `consumer.last_poll_read_partition_eof?`
 * [Racecar::Config] Remove `offset_retention_time`, `connect_timeout` and `offset_commit_threshold`
 * [Racecar::Config] Pass config to `rdkafka-ruby` via `producer` and `consumer`
 * [Racecar::Config] Replace `max_fetch_queue_size` with `min_message_queue_size`
 * [Racecar::Config] Add `synchronous_commits` to control blocking of `consumer.commit` (default `false`)
-* [Racecar::Config] Add `raise_on_partition_eof` to raise a `Rdkafka::RdkafkaError` when reaching the current end of the partition. Will only be re-raised after new messages arrive. (default `false`)
 * [Racecar::Config] Add `security_protocol` to control protocol between client and broker
 * [Racecar::Config] SSL configuration via `ssl_ca_location`, `ssl_crl_location`, `ssl_keystore_location` and `ssl_keystore_password`
 * [Racecar::Config] SASL configuration via `sasl_mechanism`, `sasl_kerberos_service_name`, `sasl_kerberos_principal`, `sasl_kerberos_kinit_cmd`, `sasl_kerberos_keytab`, `sasl_kerberos_min_time_before_relogin`, `sasl_username` and `sasl_password`